chore: refactor event types
