+++ src/core/events/event-types/network-deployed-event.ts
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {SoloEventType} from './event-types.js';
+import {SoloEvent} from './solo-event.js';
+
+export class NetworkDeployedEvent extends SoloEvent {
+  public constructor(public readonly deployment: string) {
+    super(SoloEventType.NetworkDeployed);
+  }
+}
