+++ src/core/events/event-types/nodes-started-event.ts
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {SoloEventType} from './event-types.js';
+import {SoloEvent} from './solo-event.js';
+
+export class NodesStartedEvent extends SoloEvent {
+  public constructor(public readonly deployment: string) {
+    super(SoloEventType.NodesStarted);
+  }
+}
