+++ src/core/events/event-types/solo-event.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type SoloEventType} from './event-types.js';
+
+export abstract class SoloEvent {
+  public constructor(public readonly type: SoloEventType) {}
+}
