+++ test/unit/core/events/solo-event-bus.test.ts
@@ -5,7 +5,9 @@ import {expect} from 'chai';
 import {describe, it, beforeEach} from 'mocha';
 import {type SoloEventBus} from '../../../../src/core/events/solo-event-bus.js';
 import {DefaultSoloEventBus} from '../../../../src/core/events/default-solo-event-bus.js';
-import {SoloEventType, NetworkDeployedEvent, MirrorNodeDeployedEvent} from '../../../../src/core/events/event-types.js';
+import {SoloEventType} from '../../../../src/core/events/event-types/event-types.js';
+import {NetworkDeployedEvent} from '../../../../src/core/events/event-types/network-deployed-event.js';
+import {MirrorNodeDeployedEvent} from '../../../../src/core/events/event-types/mirror-node-deployed-event.js';
 import {container} from 'tsyringe-neo';
 import {InjectTokens} from '../../../../src/core/dependency-injection/inject-tokens.js';
 import {type SoloLogger} from '../../../../src/core/logging/solo-logger.js';
